+++ data_prep/R/add_geography_cols.R
@@ -5,11 +5,15 @@ add_geography_cols <- function(arrw, year){
 
   # get code_muni from weighting area
   col <- case_when(year == 1970 ~ 'code_muni',
+                   #   year == 1980 ~ 'V5', 1 digito (mesmo V2 V3 V4 V5 ficam com 5 digitos)
+                   year == 1991 ~ 'code_muni',
                    year == 2010 ~ 'V0011',
                    year == 2000 ~ 'AREAP')
 
+  if(year %in% c(2000, 2010)){
   arrw <- mutate(arrw,
-                 code_muni := substring(get(col) , 1, 7))
+                 code_muni := as.character(substring(get(col) , 1, 7)))
+  }
 
   # weighting area
   if(year %in% c(2000, 2010)){
@@ -19,38 +23,41 @@ add_geography_cols <- function(arrw, year){
 
   # state name
   col <- case_when(year == 1970 ~ 'code_state',
+                   year == 1991 ~ 'V1101',
                    year == 2010 ~ 'V0001',
                    year == 2000 ~ 'V0102')
   arrw <- mutate(arrw,
-                 code_state = get(col),
+                 code_state = get(col))
+
+  arrw <- mutate(arrw,
                  name_state = case_when(
-                   code_state == 11 ~'Rondônia',
-                   code_state == 12 ~'Acre',
-                   code_state == 13 ~'Amazonas',
-                   code_state == 14 ~'Roraima',
-                   code_state == 15 ~'Pará',
-                   code_state == 16 ~'Amapá',
-                   code_state == 17 ~'Tocantins',
-                   code_state == 21 ~'Maranhão',
-                   code_state == 22 ~'Piauí',
-                   code_state == 23 ~'Ceará',
-                   code_state == 24 ~'Rio Grande do Norte',
-                   code_state == 25 ~'Paraíba',
-                   code_state == 26 ~'Pernambuco',
-                   code_state == 27 ~'Alagoas',
-                   code_state == 28 ~'Sergipe',
-                   code_state == 29 ~'Bahia',
-                   code_state == 31 ~'Minas Gerais',
-                   code_state == 32 ~'Espírito Santo',
-                   code_state == 33 ~'Rio de Janeiro',
-                   code_state == 35 ~'São Paulo',
-                   code_state == 41 ~'Paraná',
-                   code_state == 42 ~'Santa Catarina',
-                   code_state == 43 ~'Rio Grande do Sul',
-                   code_state == 50 ~'Mato Grosso do Sul',
-                   code_state == 51 ~'Mato Grosso',
-                   code_state == 52 ~'Goiás',
-                   code_state == 53 ~'Distrito Federal'))
+                   code_state == '11' ~'Rondônia',
+                   code_state == '12' ~'Acre',
+                   code_state == '13' ~'Amazonas',
+                   code_state == '14' ~'Roraima',
+                   code_state == '15' ~'Pará',
+                   code_state == '16' ~'Amapá',
+                   code_state == '17' ~'Tocantins',
+                   code_state == '21' ~'Maranhão',
+                   code_state == '22' ~'Piauí',
+                   code_state == '23' ~'Ceará',
+                   code_state == '24' ~'Rio Grande do Norte',
+                   code_state == '25' ~'Paraíba',
+                   code_state == '26' ~'Pernambuco',
+                   code_state == '27' ~'Alagoas',
+                   code_state == '28' ~'Sergipe',
+                   code_state == '29' ~'Bahia',
+                   code_state == '31' ~'Minas Gerais',
+                   code_state == '32' ~'Espírito Santo',
+                   code_state == '33' ~'Rio de Janeiro',
+                   code_state == '35' ~'São Paulo',
+                   code_state == '41' ~'Paraná',
+                   code_state == '42' ~'Santa Catarina',
+                   code_state == '43' ~'Rio Grande do Sul',
+                   code_state == '50' ~'Mato Grosso do Sul',
+                   code_state == '51' ~'Mato Grosso',
+                   code_state == '52' ~'Goiás',
+                   code_state == '53' ~'Distrito Federal'))
   # abbrev name
   arrw <- mutate(arrw,
                  abbrev_state = case_when(
@@ -92,14 +99,31 @@ add_geography_cols <- function(arrw, year){
                    code_region == 4 ~'Sul',
                    code_region == 5 ~'Centro-oeste'))
 
-  # reoder columns
-  if (year %in% c(2000, 2010)) {
-    arrw <- relocate(arrw, c(code_muni, code_state, abbrev_state, name_state, code_region, name_region, code_weighting))
+  # other regions
+  if(year %in% c(1991)){
+    arrw <- mutate(arrw,
+                   code_meso = V7001,
+                   code_micro = V7002,
+                   code_metro = V7003)
   }
 
-    if(year %in% c(1970)) {
+
+
+
+  ## reoder columns
+
+  if (year %in% c(2000, 2010)) {
+      arrw <- relocate(arrw, c(code_muni, code_state, abbrev_state, name_state, code_region, name_region, code_weighting))
+      }
+
+  if (year %in% c(1991)) {
+    arrw <- relocate(arrw, c(code_muni, code_state, abbrev_state, name_state, code_region, name_region, code_meso, code_micro, code_metro))
+  }
+
+  if (year %in% c(1970)) {
       arrw <- relocate(arrw, c(code_muni, code_muni_1970, code_state, abbrev_state, name_state, code_region, name_region))
-    }
+      }
+
 
 
       return(arrw)